Former-commit-id: c1aa6fd0e6a14cc62f9f4ad927bde478c8bea162

+++ rs-dam/dam/dam-service/src/main/java/fr/cnes/regards/modules/dam/service/entities/AbstractEntityService.java
@@ -368,8 +368,13 @@ public abstract class AbstractEntityService<F extends EntityFeature, U extends A
         // Set IpId
         if (entity.getIpId() == null) {
             entity.setIpId(new OaisUniformResourceName(OAISIdentifier.AIP, EntityType.valueOf(entity.getType()),
-                    runtimeTenantResolver.getTenant(), UUID.randomUUID(), 1));
+                    runtimeTenantResolver.getTenant(), UUID.randomUUID(), 1, null, null));
         }
+
+        // As long as their is no way to create new entity version thanks to dam,
+        // we set last flag and virtualId unconditionally
+        entity.setLast(true);
+        entity.setVirtualId();
 
         // IpIds of entities that will need an AMQP event publishing
         Set<UniformResourceName> updatedIpIds = new HashSet<>();
@@ -621,7 +626,8 @@ public abstract class AbstractEntityService<F extends EntityFeature, U extends A
         datasets.remove(toDelete);
         // Remove relate files
         for (Map.Entry<DataType, DataFile> entry : toDelete.getFiles().entries()) {
-            if (localStorageService.isFileLocallyStored(toDelete, entry.getValue())) {
+            if ((entry != null) && (entry.getValue() != null)
+                    && localStorageService.isFileLocallyStored(toDelete, entry.getValue())) {
                 localStorageService.removeFile(toDelete, entry.getValue());
             }
         }
